tx20240803

+++ storagemarket/direct_deals_provider.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	addr "github.com/filecoin-project/go-address"
+	miner13 "github.com/filecoin-project/go-state-types/builtin/v13/miner"
+	"io"
 	"os"
 	"runtime/debug"
 	"strings"
@@ -327,34 +330,54 @@ func (ddp *DirectDealsProvider) execDeal(ctx context.Context, entry *smtypes.Dir
 	log.Infow("processing direct deal", "uuid", dealUuid, "checkpoint", entry.Checkpoint, "piececid", entry.PieceCID, "filepath", entry.InboundFilePath, "clientAddr", entry.Client, "allocationId", entry.AllocationID)
 	ddp.dealLogger.Infow(dealUuid, "deal execution initiated", "deal state", entry)
 
-	if entry.Checkpoint <= dealcheckpoints.Accepted { // before commp
+	var isTxCar = false
+	txCarInfo, err := ParseTxCarInfo(entry.InboundFilePath)
+	if err == nil {
+		isTxCar = true
+	}
 
-		// os stat
-		fstat, err := os.Stat(entry.InboundFilePath)
-		if err != nil {
-			return &dealMakingError{
-				error: fmt.Errorf("failed to open file '%s': %w", entry.InboundFilePath, err),
-				retry: smtypes.DealRetryFatal,
+	if entry.Checkpoint <= dealcheckpoints.Accepted { // before commp
+		var carSize int64
+		if isTxCar {
+			carSize = txCarInfo.CarSize
+		} else {
+			// os stat
+			fstat, err := os.Stat(entry.InboundFilePath)
+			if err != nil {
+				return &dealMakingError{
+					error: fmt.Errorf("failed to open file '%s': %w", entry.InboundFilePath, err),
+					retry: smtypes.DealRetryFatal,
+				}
 			}
+			carSize = fstat.Size()
 		}
 
-		ddp.dealLogger.Infow(dealUuid, "size of deal", "filepath", entry.InboundFilePath, "size", fstat.Size())
+		ddp.dealLogger.Infow(dealUuid, "size of deal", "filepath", entry.InboundFilePath, "size", carSize)
 		ddp.dealLogger.Infow(dealUuid, "generating commp")
 
-		// TODO: should we be passing pieceSize here ??!?
-		pieceSize := abi.UnpaddedPieceSize(fstat.Size())
+		var generatedPieceInfo *abi.PieceInfo
+		if isTxCar {
+			generatedPieceInfo = &abi.PieceInfo{
+				Size:     abi.PaddedPieceSize(txCarInfo.PieceSize),
+				PieceCID: txCarInfo.PieceCid,
+			}
+		} else {
+			// TODO: should we be passing PieceSize here ??!?
+			pieceSize := abi.UnpaddedPieceSize(carSize)
 
-		generatedPieceInfo, dmErr := generatePieceCommitment(ctx, ddp.commpCalc, ddp.commpThrottle, entry.InboundFilePath, pieceSize.Padded(), ddp.config.RemoteCommp)
-		if dmErr != nil {
-			return &dealMakingError{
-				retry: types.DealRetryManual,
-				error: fmt.Errorf("failed to generate commp: %w", dmErr),
+			var dmErr error
+			generatedPieceInfo, dmErr = generatePieceCommitment(ctx, ddp.commpCalc, ddp.commpThrottle, entry.InboundFilePath, pieceSize.Padded(), ddp.config.RemoteCommp)
+			if dmErr != nil {
+				return &dealMakingError{
+					retry: types.DealRetryManual,
+					error: fmt.Errorf("failed to generate commp: %w", dmErr),
+				}
 			}
 		}
 
-		entry.InboundFileSize = fstat.Size()
+		entry.InboundFileSize = carSize
 
-		log.Infow("direct deal details", "filepath", entry.InboundFilePath, "supplied-piececid", entry.PieceCID, "calculated-piececid", generatedPieceInfo.PieceCID, "calculated-piecesize", generatedPieceInfo.Size, "os stat size", fstat.Size())
+		log.Infow("direct deal details", "filepath", entry.InboundFilePath, "supplied-piececid", entry.PieceCID, "calculated-piececid", generatedPieceInfo.PieceCID, "calculated-piecesize", generatedPieceInfo.Size, "os stat size", carSize)
 
 		if !entry.PieceCID.Equals(generatedPieceInfo.PieceCID) {
 			return &dealMakingError{
@@ -362,6 +385,7 @@ func (ddp *DirectDealsProvider) execDeal(ctx context.Context, entry *smtypes.Dir
 				error: fmt.Errorf("commP expected=%s, actual=%s: %w", entry.PieceCID, generatedPieceInfo.PieceCID, ErrCommpMismatch),
 			}
 		}
+
 		ddp.dealLogger.Infow(dealUuid, "completed generating commp")
 
 		entry.PieceSize = generatedPieceInfo.Size
@@ -409,7 +433,12 @@ func (ddp *DirectDealsProvider) execDeal(ctx context.Context, entry *smtypes.Dir
 					Client: abi.ActorID(clientId),
 					ID:     verifreg13types.AllocationId(entry.AllocationID),
 				},
-				Notify: nil,
+				Notify: []miner13.DataActivationNotification{
+					{
+						Address: addr.Undef,
+						Payload: []byte(entry.InboundFilePath),
+					},
+				},
 			},
 
 			// Best-effort deal asks
@@ -418,11 +447,16 @@ func (ddp *DirectDealsProvider) execDeal(ctx context.Context, entry *smtypes.Dir
 
 		// Open a reader over the piece data
 		ddp.dealLogger.Infow(dealUuid, "opening reader over piece data", "filepath", entry.InboundFilePath)
-		paddedReader, err := openReader(entry.InboundFilePath, entry.PieceSize.Unpadded())
-		if err != nil {
-			return &dealMakingError{
-				retry: types.DealRetryFatal,
-				error: fmt.Errorf("failed to read piece data: %w", err),
+		var paddedReader io.ReadCloser
+		if isTxCar {
+			paddedReader = io.NopCloser(strings.NewReader(entry.InboundFilePath))
+		} else {
+			paddedReader, err = openReader(entry.InboundFilePath, entry.PieceSize.Unpadded())
+			if err != nil {
+				return &dealMakingError{
+					retry: types.DealRetryFatal,
+					error: fmt.Errorf("failed to read piece data: %w", err),
+				}
 			}
 		}
 
@@ -449,8 +483,10 @@ func (ddp *DirectDealsProvider) execDeal(ctx context.Context, entry *smtypes.Dir
 
 		// The deal has been handed off to the sealer, so we can remove the inbound file
 		if entry.CleanupData {
-			_ = os.Remove(entry.InboundFilePath)
-			ddp.dealLogger.Infow(dealUuid, "removed piece data from disk as deal has been added to a sector", "path", entry.InboundFilePath)
+			if !isTxCar {
+				_ = os.Remove(entry.InboundFilePath)
+				ddp.dealLogger.Infow(dealUuid, "removed piece data from disk as deal has been added to a sector", "path", entry.InboundFilePath)
+			}
 		}
 	}
 
@@ -673,6 +709,8 @@ func (ddp *DirectDealsProvider) FailPausedDeal(ctx context.Context, id uuid.UUID
 func (ddp *DirectDealsProvider) indexAndAnnounce(ctx context.Context, entry *smtypes.DirectDeal) *dealMakingError {
 	// If this is Curio sealer then we should wait till sector finishes sealing
 	if ddp.config.Curio {
+		log.Infow("----indexAndAnnounce", "uuid", entry.SectorID, "checkpoint", entry.Checkpoint, "piececid", entry.PieceCID, "filepath", entry.InboundFilePath, "clientAddr", entry.Client, "allocationId", entry.AllocationID)
+
 		// Wait for sector to finish sealing
 		err := ddp.trackCurioSealing(entry.SectorID)
 		if err != nil {
